Refactor: SessionManager auth 패키지로 이동
